To_Delete_Later/aria/requirements.txt
# ARIA - AI Home Assistant Dependencies

# Core
python-dotenv>=1.0.0
requests>=2.31.0

# Google Gemini
google-genai>=1.0.0

# Audio I/O
pyaudio>=0.2.14
sounddevice>=0.4.6
numpy>=1.24.0

# Speech-to-Text (local, fast)
faster-whisper>=1.0.0

# Text-to-Speech
edge-tts>=6.1.0
pygame>=2.5.0

# Wake word detection (local)
openwakeword>=0.6.0
onnxruntime>=1.16.0

# RAG - Qdrant
qdrant-client>=1.7.0

# Home automation
yeelight>=0.7.0

# YouTube
yt-dlp>=2024.1.0

# Weather
python-weather>=2.0.0

# Email
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0
